add clerk
