+++ src/pages/_app.tsx
@@ -1,9 +1,8 @@
 import { type Session } from "next-auth";
-import { SessionProvider } from "next-auth/react";
 import { type AppType } from "next/app";
 import { Inter as FontSans } from "next/font/google";
-import { Navbar } from "~/components/navbar";
 import "~/styles/globals.css";
+import { ClerkProvider } from "@clerk/nextjs";
 
 import { api } from "~/utils/api";
 import { cn } from "~/utils/utils";
@@ -18,7 +17,23 @@ const MyApp: AppType<{ session: Session | null }> = ({
   pageProps: { session, ...pageProps },
 }) => {
   return (
-    <SessionProvider session={session}>
+    <ClerkProvider
+      appearance={{
+        variables: { colorPrimary: "#000000" },
+        elements: {
+          formButtonPrimary:
+            "bg-black border border-black border-solid hover:bg-white hover:text-black",
+          socialButtonsBlockButton:
+            "bg-white border-gray-200 hover:bg-transparent hover:border-black text-gray-600 hover:text-black",
+          socialButtonsBlockButtonText: "font-semibold",
+          formButtonReset:
+            "bg-white border border-solid border-gray-200 hover:bg-transparent hover:border-black text-gray-500 hover:text-black",
+          membersPageInviteButton:
+            "bg-black border border-black border-solid hover:bg-white hover:text-black",
+          card: "bg-[#fafafa]",
+        },
+      }}
+    >
       <Component
         className={cn(
           "min-h-screen bg-background font-sans antialiased",
@@ -26,7 +41,7 @@ const MyApp: AppType<{ session: Session | null }> = ({
         )}
         {...pageProps}
       />
-    </SessionProvider>
+    </ClerkProvider>
   );
 };
