+++ src/pages/addItem.tsx
@@ -1,18 +1,16 @@
 import React, { useEffect, useState } from "react";
-import Link from "next/link";
 import { api } from "~/utils/api";
 import { Button } from "~/components/ui/button";
-import { useSession } from "next-auth/react";
 import router from "next/router";
+import { useUser } from "@clerk/nextjs";
 
 export default function AddItem() {
   const utils = api.useUtils();
-  const { data: sessionData } = useSession();
-  const res = sessionData as unknown as { id: string };
   const [uid, setUid] = useState<string>("");
+  const { isLoaded, user } = useUser();
+
   useEffect(() => {
-    if (!res) router.push("/");
-    if (res) setUid(res.id);
+    if (isLoaded && user) setUid(user.id);
   });
 
   const dbItem = api.cashier.createItem.useMutation({
