+++ src/pages/reports.tsx
@@ -1,6 +1,6 @@
 import EarningsReport from "./dashboard/reports";
 
-export default function Dashboard() {
+export default function Reports() {
   return (
     <div>
       <EarningsReport todaysEarnings={2500} totalEarnings={35000} />
